Refactored and adopting pytest. Added License
