+++ src/sdk/python/rtdip_sdk/data_models/usage/Usage.py
@@ -1,0 +1,23 @@
+# Copyright 2022 RTDIP
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from pydantic import BaseModel
+
+
+class Usage(BaseModel):
+    uid: str  # A unique identifier tied to a timeseries of data.
+    series_id: str  # Identifier for a particular timeseries set
+    timestamp: int  # Creation time. Always UTC. Seconds since EPOCH
+    interval_timestamp: int  # The timestamp for the interval. Always UTC. Seconds since EPOCH.
+    value: float  # The actual value of the measurement
