+++ src/sdk/python/rtdip_sdk/data_models/weather_factory/weather_object_factory.py
@@ -1,0 +1,29 @@
+
+# Copyright 2022 RTDIP
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from abc import ABCMeta, abstractmethod
+
+
+class WeatherObjectFactory(metaclass=ABCMeta):
+    @staticmethod
+    @abstractmethod
+    def create_object(self):
+        "Creates an Instance of the Weather Object"
+
+
+
+
+
+
