--- src/staging/data_models/requirements.txt
@@ -1,36 +1,0 @@
-# Innowatts CONFIDENTIAL
-# Copyright
-# Innowatts Limited All Rights Reserved.
-#
-# The source code contained or described  herein and all documents related to
-# the source code ("Material") are owned by  Innowatts Limited or its suppliers
-# or licensors.
-#
-# Title to the Material remains with Innowatts Limited or its suppliers and
-# licensors.
-#
-# The Material contains trade secrets and proprietary and confidential
-# information of Innowatts or its suppliers and licensors.
-#
-# The Material is protected by worldwide copyright and trade secret laws and
-# treaty provisions.
-#
-# No part of the Material may be used, copied, reproduced, modified, published,
-# uploaded, posted, transmitted, distributed, or disclosed in any way
-# without  Innowatts's prior express written permission.
-#
-# No license under any patent, copyright, trade secret or other IP right is
-# granted to or conferred upon you by disclosure or delivery of the Materials,
-# either expressly, by implication, inducement, estoppel or otherwise.
-# Any license under such IP rights must be express and approved by Innowatts
-# in writing.
-
-# To find out which version is installed  type pip show <package>
-# this file will be utilised when generating distributions later
-pyramid==1.10.4
-pyyaml==5.2
-python-json-logger==0.1.11
-dataclasses-json==0.5.7
-tzlocal==4.3
-pandas==1.5.3
-
