--- src/staging/data_models/src/data_models/Constants/weather_constants.py
@@ -1,2 +1,0 @@
-AtmosphericG215minForecastV1 = 'AtmosphericG215minForecastV1'
-version = 'version'
