--- src/staging/data_models/src/data_models/usage/Version.py
@@ -1,3 +1,0 @@
-
-class Version:
-    Version_0_0_1: str = 'v0.0.1'
