--- src/staging/data_models/src/data_models/usage/pjm_voltage_lvl.py
@@ -1,10 +1,0 @@
-from enum import Enum
-
-
-class PjmVoltagelvl(Enum):
-    transmission = 0
-    sub_transmission = 1
-    primary_substation = 2
-    primary = 3
-    secondary = 4
-    unmetered = 5
