--- src/staging/data_models/src/data_models/weather_factory/weather_object_factory.py
@@ -1,14 +1,0 @@
-from abc import ABCMeta, abstractmethod
-
-
-class WeatherObjectFactory(metaclass=ABCMeta):
-    @staticmethod
-    @abstractmethod
-    def create_object(self):
-        "Creates an Instance of the Weather Object"
-
-
-
-
-
-
